Send periodic updates of the entities in STARSr
We send a message on startup to the specified
port and address (default 127.0.0.1:31415, use
--piloting switch on server.py to change).

The first message is really big and is sent over TCP.

Further messages of values that change over the
operation of the program will be sent over UDP to the
same port.

Right now it's a very simple dictionary of lists,
sent with msgpack. Use msgpack on the receiving side
to decode a JSON object.

Example receiving server in test_scripts.